moves sources to its own package
