+++ sqlflow/sources/webhook.py
@@ -1,10 +1,10 @@
 import hmac
 import hashlib
 import logging
-import typing
-from abc import ABC, abstractmethod
-from typing import Iterator
 from threading import Thread, Lock
+import typing
+
+from .base import Source, Message
 
 from fastapi import FastAPI, Request, HTTPException
 from opentelemetry import metrics
@@ -13,14 +13,9 @@ from starlette.middleware.base import BaseHTTPMiddleware
 from queue import Queue, Empty
 import uvicorn
 
-from websockets.sync.client import connect
-
-from confluent_kafka import KafkaError, Consumer
-
-from sqlflow import config
-
 logger = logging.getLogger(__name__)
 meter = metrics.get_meter('sqlflow.sources.http')
+
 
 REQUEST_COUNT = meter.create_counter(
     "webhook_requests_total",
@@ -33,105 +28,6 @@ REQUEST_DURATION = meter.create_histogram(
     "seconds",
     "Duration of requests to the webhook source in seconds",
 )
-
-class Message:
-    def __init__(self, value: bytes):
-        self._value = value
-
-    def value(self) -> bytes:
-        return self._value
-
-
-class Source(ABC):
-    """
-    Abstract base class for sources.
-
-    Sources read bytes from some external source and return them as messages.
-    """
-    @abstractmethod
-    def close(self):
-        pass
-
-    @abstractmethod
-    def commit(self):
-        pass
-
-    @abstractmethod
-    def stream(self) -> Iterator[Message | None]:
-        pass
-
-    @abstractmethod
-    def start(self):
-        pass
-
-
-class SourceException(Exception):
-    pass
-
-
-class KafkaSource(Source):
-    def __init__(self,
-                 consumer,
-                 topics,
-                 async_commit=False,
-                 read_timeout=5.0):
-        self._consumer = consumer
-        self._async_commit = async_commit
-        self._read_timeout = read_timeout
-        self._topics = topics
-
-    def close(self):
-        return self._consumer.close()
-
-    def commit(self):
-        return self._consumer.commit(
-            asynchronous=self._async_commit,
-        )
-
-    def stream(self) -> typing.Iterable[Message | None]:
-        while True:
-            msg = self._consumer.poll(timeout=self._read_timeout)
-
-            if msg is None:
-                yield None
-            elif not msg.error():
-                yield Message(msg.value())
-            elif msg.error().code() == KafkaError._PARTITION_EOF:
-                logger.warning(
-                    '%% %s [%d] reached end at offset %d'.format(
-                        msg.topic(),
-                        msg.partition(),
-                        msg.offset(),
-                    )
-                )
-                yield None
-            else:
-                # unknown error raise to caller
-                raise SourceException(msg.error())
-
-    def start(self):
-        self._consumer.subscribe(self._topics)
-
-
-class WebsocketSource(Source):
-    def __init__(self, uri):
-        self._uri = uri
-
-    def start(self):
-        pass
-
-    def commit(self):
-        pass
-
-    def close(self):
-        pass
-
-    def stream(self) -> Message | None:
-        logger.info("connecting to websocket: {}".format(self._uri))
-        with connect(self._uri) as websocket:
-            while True:
-                for msg in websocket.recv_streaming(decode=False):
-                    yield Message(msg)
 
 
 class HMACConfig:
@@ -224,36 +120,4 @@ class WebhookSource(Source):
     def close(self):
         # Close the source (if needed)
         pass
-
-
-def new_source_from_conf(source_conf: config.Source):
-    if source_conf.type == 'kafka':
-        kconf = {
-            'bootstrap.servers': ','.join(source_conf.kafka.brokers),
-            'group.id': source_conf.kafka.group_id,
-            'auto.offset.reset': source_conf.kafka.auto_offset_reset,
-            'enable.auto.commit': False,
-        }
-
-        consumer = Consumer(kconf)
-
-        return KafkaSource(
-            consumer=consumer,
-            topics=source_conf.kafka.topics,
-        )
-    elif source_conf.type == 'websocket':
-        return WebsocketSource(
-            uri=source_conf.websocket.uri,
-        )
-    elif source_conf.type == 'webhook':
-        hmac_config = None
-        if source_conf.webhook.signature_type == 'hmac' and source_conf.webhook.hmac:
-            hmac_config = HMACConfig(
-                header=source_conf.webhook.hmac.header,
-                sig_key=source_conf.webhook.hmac.sig_key,
-                secret=source_conf.webhook.hmac.secret,
-            )
-        return WebhookSource(hmac_config=hmac_config)
-
-    raise NotImplementedError('unsupported source type: {}'.format(source_conf.type))
